backend/requirements.txt
flask
pytest
psycopg2-binary
simplejson
python-dotenv
requests
